+++ vite.config.js
@@ -1,43 +1,33 @@
 /* global __dirname */
 import { resolve } from 'path';
 import { defineConfig } from 'vite';
-import { configDefaults } from 'vitest/config';
 // import babel from 'vite-plugin-babel';
 
 import banner from 'vite-plugin-banner';
 import pkg from './package.json';
 
+const origName = pkg.name;
+
 const bannerText = `/*!
-* Zero2One v${pkg.version}
+* ${origName} v${pkg.version}
 * ${pkg.homepage}
 */`;
+
+const lookupNames = {
+  es: `${origName}.js`,
+  cjs: `${origName}.cjs.js`,
+  umd: `${origName}.umd.js`,
+};
 
 export default defineConfig({
   build: {
     target: 'es2015', // esnext
     lib: {
-      entry: resolve(__dirname, 'src/zero2one.js'),
+      entry: resolve(__dirname, 'lib/zero2one.js'),
       name: 'Zero2One',
-      fileName: (format) => {
-        format = 'es' === format ? '' : `.${format}`;
-        return `zero2one${format}.js`;
-      },
-    },
-  },
-
-  test: {
-    globals: true,
-    environment: 'jsdom',
-    coverage: {
-      provider: 'v8', // 'istanbul' or 'v8'
-      exclude: [
-        ...configDefaults.exclude,
-        '_notes/**',
-        'config/**',
-        'docs/**',
-        'test/*.bench.*',
-        '.eslintrc.js',
-      ],
+      formats: ['es', 'cjs', 'umd'],
+      // @ts-expect-error: i only need 'es', 'cjs', 'umd'
+      fileName: (format) => lookupNames[format],
     },
   },