+++ vitest.config.js
@@ -1,0 +1,20 @@
+import { defineConfig } from 'vitest/config';
+import { configDefaults } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    coverage: {
+      provider: 'v8', // 'istanbul' or 'v8'
+      exclude: [
+        ...configDefaults.exclude,
+        '_notes/**',
+        'config/**',
+        'docs/**',
+        'test/*.bench.*',
+        '.eslintrc.js',
+      ],
+    },
+  },
+});